requirements.txt
dash==3.2.0
dash-bootstrap-components==2.0.3
pandas==2.3.1
numpy==2.3.0
plotly==6.2.0
matplotlib==3.10.5
dtw-python==1.5.3
pyarrow==21.0.0

